refactor : car, racing package로 domain package 구성
