+++ src/test/kotlin/next/step/racing/domain/CarNameTest.kt
@@ -1,7 +1,8 @@
-package next.step.racing.domain.car
+package next.step.racing.domain
 
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.DescribeSpec
+import next.step.racing.domain.car.CarName
 
 class CarNameTest : DescribeSpec({
