+++ src/test/kotlin/next/step/racing/domain/CarPositionTest.kt
@@ -1,8 +1,9 @@
-package next.step.racing.domain.car
+package next.step.racing.domain
 
 import io.kotest.core.spec.style.DescribeSpec
 import io.kotest.datatest.withData
 import io.kotest.matchers.shouldBe
+import next.step.racing.domain.car.CarPosition
 
 class CarPositionTest : DescribeSpec({
