+++ src/test/kotlin/next/step/racing/domain/CarTest.kt
@@ -1,8 +1,11 @@
-package next.step.racing.domain.car
+package next.step.racing.domain
 
 import io.kotest.core.spec.style.DescribeSpec
 import io.kotest.datatest.withData
 import io.kotest.matchers.shouldBe
+import next.step.racing.domain.car.Car
+import next.step.racing.domain.car.CarName
+import next.step.racing.domain.car.CarPosition
 import next.step.racing.util.DrivingStrategy
 
 class CarTest : DescribeSpec({
